+++ apps/sb-vite/src/main.tsx
@@ -1,10 +1,13 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
+import '@tamagui/core/reset.css'
+import '@fontsource/inter/400.css'
+import '@fontsource/inter/700.css'
+import 'raf/polyfill'
 import App from './App.tsx'
-import './index.css'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
+  </React.StrictMode>
 )