+++ packages/date/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @tamagui-extras/date
 
+## 1.9.11
+
+### Patch Changes
+
+- update dependencies
+- Updated dependencies
+  - @tamagui-extras/core@1.8.6
+  - @tamagui-extras/form@1.13.4
+
 ## 1.9.10
 
 ### Patch Changes