+++ packages/link/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tamagui-extras/link
 
+## 1.8.6
+
+### Patch Changes
+
+- update dependencies
+
 ## 1.8.5
 
 ### Patch Changes