+++ packages/video/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tamagui-extras/video
 
+## 1.9.6
+
+### Patch Changes
+
+- update dependencies
+
 ## 1.9.5
 
 ### Patch Changes